# src/main/resources/api-specs/openapi.v1.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/api/v1/store":{"post":{"tags":["store-controller"],"operationId":"store","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"},"request":{"$ref":"#/components/schemas/UrlStoreRequest"}}}},"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"},"request":{"$ref":"#/components/schemas/UrlStoreRequest"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/chat":{"post":{"tags":["chat-controller"],"operationId":"chat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatMessageResponse"}}}}}}},"/api/v1/chat/{conversationId}":{"get":{"tags":["chat-controller"],"operationId":"getMessages","parameters":[{"name":"conversationId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Message"}}}}}}},"post":{"tags":["chat-controller"],"operationId":"chat_1","parameters":[{"name":"conversationId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatMessageResponse"}}}}}}},"/ui":{"get":{"tags":["ai-test-controller"],"operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/ui/prompt":{"get":{"tags":["ai-test-controller"],"operationId":"prompt","parameters":[{"name":"text","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UrlStoreRequest":{"type":"object","properties":{"url":{"type":"string"}}},"ChatMessageRequest":{"type":"object","properties":{"message":{"type":"string"}}},"ChatMessageResponse":{"type":"object","properties":{"conversationId":{"type":"string"},"message":{"$ref":"#/components/schemas/Message"}}},"Message":{"type":"object","properties":{"messageType":{"type":"string","enum":["USER","ASSISTANT","SYSTEM","TOOL"]},"metadata":{"type":"object","additionalProperties":{}},"text":{"type":"string"}}}}}}